requirements.txt
absl-py==1.4.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
filelock==3.10.0
flatbuffers==23.3.3
fonttools==4.39.2
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
identify==2.5.21
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
nodeenv==1.7.0
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
platformdirs==3.1.1
pre-commit==3.2.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sklearn==0.0.post1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
